Add files via upload
This project is a comprehensive Power BI report created using the Sample EU Superstore dataset. The goal of this project was to demonstrate proficiency in Power BI by focusing on key areas such as structured data model with relationships and calculated columns/measures and implemented  complex DAX measures to showcase analytical capabilities. The report tells a coherent story based on the analysis, highlighting key business insights for decision-makers.